[ios] Fix touch handling in groups
The problem was in native code - it wasn't caching path created by
internal elements.
Another thing is that implementation of `hitTest` should return
"farthest descendant of the receiver in the view hierarchy" which may
result in incorrect behavior when both group and internal element are
set up to handle touches. If touch is landing on internal view - then it
should handle touch, wrapping group otherwise.

https://developer.apple.com/documentation/uikit/uiview/1622469-hittest?language=objc